# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Prompt-to-Form API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: "#/components/schemas/UpdateMeBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /forms:
    get:
      operationId: listForms
      tags: [forms]
      summary: List user's forms
      responses:
        "200":
          description: List of forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
        "401":
          description: Unauthorized

    post:
      operationId: createForm
      tags: [forms]
      summary: Create a new form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormBody"
      responses:
        "201":
          description: Form created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormWithFields"
        "401":
          description: Unauthorized
  /forms/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getForm
      tags: [forms]
      summary: Get form details
      responses:
        "200":
          description: Form details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormWithFields"
        "404":
          description: Not found

    patch:
      operationId: updateForm
      tags: [forms]
      summary: Update form settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFormBody"
      responses:
        "200":
          description: Updated form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
        "404":
          description: Not found

    delete:
      operationId: deleteForm
      tags: [forms]
      summary: Delete a form
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /forms/{id}/duplicate:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: duplicateForm
      tags: [forms]
      summary: Duplicate an existing form
      responses:
        "201":
          description: Duplicated form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
        "404":
          description: Not found
  /forms/{id}/publish:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: publishForm
      tags: [forms]
      summary: Publish form and generate translations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishFormBody"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishFormResult"
        "404":
          description: Not found
  /forms/{id}/sentiment-summary:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: generateFormSentimentSummary
      tags: [forms]
      summary: Generate AI sentiment analysis summary for a form
      responses:
        "200":
          description: Sentiment summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
        "404":
          description: Not found
  /forms/{id}/submissions:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List all submissions for a form
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
...
  /forms/{id}/translations:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listFormTranslations
      tags: [translations]
      summary: List all translations for a form
      responses:
        "200":
          description: List of translations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormTranslation"
...
  /forms/{id}/fields:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listFormFields
      tags: [fields]
      summary: List all fields for a form
      responses:
        "200":
          description: List of fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormField"

    post:
      operationId: createFormField
      tags: [fields]
      summary: Add a new field to a form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFieldBody"
      responses:
        "201":
          description: Created field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormField"
  /forms/{id}/fields/reorder:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: reorderFields
      tags: [fields]
      summary: Reorder fields in a form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderFieldsBody"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /forms/{id}/fields/{fieldId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: fieldId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateFormField
      tags: [fields]
      summary: Update a specific field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFieldBody"
      responses:
        "200":
          description: Updated field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormField"

    delete:
      operationId: deleteFormField
      tags: [fields]
      summary: Delete a field
      responses:
        "204":
          description: Deleted
...
  /public/forms/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublicForm
      tags: [forms]
      summary: Get a published form by slug
      responses:
        "200":
          description: Public form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicForm"
        "404":
          description: Not found
  /public/forms/{slug}/submit:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: submitForm
      tags: [submissions]
      summary: Submit a form response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFormBody"
      responses:
        "201":
          description: Submission recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Validation error or form closed
  /ai/generate:
    post:
      operationId: generateForm
      tags: [forms]
      summary: Generate a form structure from a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateFormBody"
      responses:
        "200":
          description: Generated form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateFormResult"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
      required: [id, email]
    Form:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        slug:
          type: string
        originalLanguage:
          type: string
        preferredLanguage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, closed]
        supportedLanguages:
          type: array
          items:
            type: string
        responseLimit:
          type: integer
          nullable: true
        closesAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, title, slug, originalLanguage, status, supportedLanguages, createdAt]
    FormField:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        orderIndex:
          type: integer
        fieldType:
          type: string
          enum: [short_text, long_text, single_choice, multi_choice, rating, date, email, phone]
        label:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        optionsJson:
          type: array
          items:
            type: string
          nullable: true
      required: [id, formId, orderIndex, fieldType, label, isRequired]
    FormTranslation:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        language:
          type: string
        translationsJson:
          type: object
          additionalProperties: true
        generatedAt:
          type: string
      required: [id, formId, language, translationsJson, generatedAt]
    Submission:
      type: object
      properties:
        id:
          type: string
        formId:
          type: string
        respondentLanguage:
          type: string
        rawResponsesJson:
          type: object
          additionalProperties: true
        translatedResponsesJson:
          type: object
          additionalProperties: true
          nullable: true
        sentimentAnalysisJson:
          type: object
          additionalProperties: true
          nullable: true
        translationStatus:
          type: string
          enum: [pending, done, skipped]
        submittedAt:
          type: string
      required: [id, formId, respondentLanguage, rawResponsesJson, translationStatus, submittedAt]
    FormWithFields:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        slug:
          type: string
        originalLanguage:
          type: string
        preferredLanguage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, closed]
        supportedLanguages:
          type: array
          items:
            type: string
        responseLimit:
          type: integer
          nullable: true
        closesAt:
          type: string
          nullable: true
        createdAt:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
      required: [id, userId, title, slug, originalLanguage, status, supportedLanguages, createdAt, fields]
    PublicForm:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        slug:
          type: string
        originalLanguage:
          type: string
        supportedLanguages:
          type: array
          items:
            type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"
        translations:
          type: array
          items:
            $ref: "#/components/schemas/FormTranslation"
      required: [id, title, slug, originalLanguage, supportedLanguages, fields, translations]
...
    UpdateMeBody:
      type: object
      properties:
        password:
          type: string
          minLength: 6
        preferredLanguage:
          type: string
          nullable: true
    CreateFormBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        originalLanguage:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/CreateFieldBody"
      required: [title, originalLanguage, fields]
    UpdateFormBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        featureImageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, closed]
        preferredLanguage:
          type: string
          nullable: true
        supportedLanguages:
          type: array
          items:
            type: string
        responseLimit:
          type: integer
          nullable: true
        closesAt:
          type: string
          nullable: true
    PublishFormBody:
      type: object
      properties:
        languages:
          type: array
          items:
            type: string
      required: [languages]
    PublishFormResult:
      type: object
      properties:
        success:
          type: boolean
        failedLanguages:
          type: integer
        form:
          $ref: "#/components/schemas/Form"
      required: [success, failedLanguages, form]
    CreateFieldBody:
      type: object
      properties:
        fieldType:
          type: string
          enum: [short_text, long_text, single_choice, multi_choice, rating, date, email, phone]
        label:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        optionsJson:
          type: array
          items:
            type: string
          nullable: true
        orderIndex:
          type: integer
      required: [fieldType, label, isRequired]
    UpdateFieldBody:
      type: object
      properties:
        label:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        optionsJson:
          type: array
          items:
            type: string
          nullable: true
        orderIndex:
          type: integer
    ReorderFieldsBody:
      type: object
      properties:
        fieldIds:
          type: array
          items:
            type: string
      required: [fieldIds]
    SubmitFormBody:
      type: object
      properties:
        respondentLanguage:
          type: string
        responses:
          type: object
          additionalProperties: true
      required: [respondentLanguage, responses]
    GenerateFormBody:
      type: object
      properties:
        prompt:
          type: string
        language:
          type: string
          description: BCP-47 language code to generate the form in (e.g. "en", "fr"). Defaults to auto-detect from prompt.
      required: [prompt]
    GenerateFormResult:
      type: object
      properties:
        form:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
              nullable: true
            featureImageUrl:
              type: string
              nullable: true
            fields:
              type: array
              items:
                $ref: "#/components/schemas/CreateFieldBody"
          required: [title, fields]
        detectedLanguage:
          type: string
      required: [form, detectedLanguage]
...